Changed file structure and moved filters as seperate component
