+++ src/Pages/SignUp.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
-import { useAuth } from "../Redux/AuthProvider"
+import { useAuth } from "../Context/AuthProvider"
 const SignUp = () => {
     const [showpasswordState, setPassState] = useState(false);
     const { signinUser } = useAuth()
