+++ src/PrivateRoute.js
@@ -1,6 +1,6 @@
 import { Route, Navigate } from "react-router-dom";
 import React from "react"
-import { useAuth } from "./Redux/AuthProvider";
+import { useAuth } from "./Context/AuthProvider";
 
 export function PrivateRoute({ path, ...props }) {
   const { isUserLogin } = useAuth();
