+++ src/components/WishlistItem.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { useCart } from "../Redux/cart-context";
+import { useCart } from "../Context/cart-context";
 import { Link } from "react-router-dom";
 export function WishlistItem({ dataset }) {
   const { dispatch } = useCart();
